Disable Rails host authorization in development
In development Rails (6+) includes a middleware that rejects a request with a 403 response if its host isn't present in the allowlist (a security feature). This prevents Parklife from working in a Rails app out of the box unless you manually add the expected Parklife base to the hosts allowlist or set it to nil to disable it - both of which aren't great because they disable the security feature whenever the development server is booted.

https://guides.rubyonrails.org/configuring.html#actiondispatch-hostauthorization

However it's safe to remove the middleware via Parklife because it won't be executed in the normal Rails development flow, only via a Parkfile when parklife/rails is required.