Apply Rails default_url_options, relative_url_root, and force_ssl settings to Parklife's base (#118)
This makes it so the default Rails production setting of `force_ssl=true` is automatically applied to Parklife's base which prevents `301` responses when running the default `bin/static-build` (#116), it also applies `default_url_options` and `relative_url_root` if they're defined. Setting a `base` in the `Parkfile` or via `--base` with the CLI will override these Rails-applied defaults.